requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
pymupdf
python-dotenv